requirements.txt
asgiref==3.2.10
astroid==2.4.2
bleach==3.1.5
Django==3.1
django-markdownx==3.0.1
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
packaging==20.4
Pillow==7.2.0
pylint==2.5.3
pyparsing==2.4.7
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
webencodings==0.5.1
wrapt==1.12.1
